update README.md and version (#194)
* update README.md and version

* move multiexp code to provider/mod.rs

* update README.md

* small edits

* small edits
